Fix a couple deprecations and update Ember CLI Babel
The `errors.unknownProperty` accessor would initialise the property
on every attempt to access. If this happens more than once during a
render cycle, newer Ember will complain since it is highly inefficient.
See https://github.com/emberjs/ember.js/issues/13948 for more details.

Also removes the use of the `getOwner` polyfill, which was added
natively to `Ember.getOwner` in Ember 2.3.

Lastly updates the dependency on `ember-cli-babel` to be in line with
newer Ember CLI requirements.